[添加模块]springboot-redis的学习,redis緩存应用3,未完成

+++ spring-boot-redis/src/main/java/com/hkz/service/UserService.java
@@ -16,6 +16,6 @@ public interface UserService {
 
     User addUser(User user);
 
-    void deleteUser(long id);
+    void deleteUser(int  id);
 
 }
